Moved Voice Recognizer to own Util Script
main.py now calls this instead, this upholds the modular framework.
